backend/requirements.txt
google-genai>=1.0.0
requests
fastapi
uvicorn
python-dotenv